refactor(payment) : 서비스 통신 클라이언트 파일명과 폴더 구조 변경
